oci-language-multiple-translation/requirements.txt
oci>=2.141.1
pyyaml